Remove (routes) from app, move components from app to components
