+++ src/app/(auth)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import "./auth.css"
-import Footer from './components/Footer'
+import Footer from '@/components/auth/Footer'
 import ReduxProvider from '../ReduxProvider'
 
 const inter = Inter({ subsets: ['latin'] })
