+++ src/app/(auth)/signup/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import AuthLayout from '../../components/AuthLayout'
+import AuthLayout from '@/components/auth/AuthLayout/AuthLayout'
 
 const SignUp = () => <AuthLayout status="signup" />
