+++ src/components/auth/AuthLayout/FormContainer/EmailTemplate.tsx
