+++ src/components/auth/Footer.tsx
